Refactor `meta_config` to `meta` Drop `base` backwards-compat wrapper
