--- ovos_config/base.py
@@ -1,3 +1,0 @@
-from ovos_config.models import *
-from ovos_utils.log import LOG
-LOG.warning("This reference is deprecated, import from ovos_config.models")
