+++ ovos_config/ovos.py
@@ -1,4 +1,4 @@
-from ovos_config.meta_config import *
+from ovos_config.meta import *
 from ovos_utils.log import LOG
 LOG.warning("This reference is deprecated, "
             "import from ovos_config.meta_config")
